requirements/dev.txt
-r base.txt
coverage==4.5.3
factory-boy==2.7.0
Faker==1.0.4
flake8==3.7.7
ipdb==0.11.0
ipython==7.3.0
mock==2.0.0
pytest==4.3.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-sugar==0.9.2
pytest-xdist==1.26.1
Sphinx==1.5.3
vcrpy==2.0.1
astroid==2.2.5
isort==4.3.20
lazy-object-proxy==1.4.1
pylint==2.3.1
typed-ast==1.4.0
